BLOCKS-019 sensing/operator renaming
