+++ playground/playground.html
@@ -19,10 +19,10 @@
   <script src="../scratch-blocks-develop/blocks_vertical/motion.js"></script>
   <script src="../scratch-blocks-develop/blocks_vertical/looks.js"></script>
   <script src="../scratch-blocks-develop/blocks_vertical/procedures.js"></script>
-  <script src="../scratch-blocks-develop/blocks_vertical/operators.js"></script>
   <script src="../scratch-blocks-develop/blocks_vertical/pen.js"></script>
+  <script src="../scratch-blocks-develop/blocks_vertical/sensing.js"></script>
   <script src="../scratch-blocks-develop/blocks_vertical/sound.js"></script>
-  <script src="../scratch-blocks-develop/blocks_vertical/pen.js"></script>
+  <script src="../scratch-blocks-develop/blocks_vertical/sensing.js"></script>
   <script src="../scratch-blocks-develop/blocks_vertical/data.js"></script>
   <script src="../scratch-blocks-develop/blocks_vertical/extensions.js"></script>
