+++ scratch-blocks-develop/tests/playground.html
@@ -19,9 +19,9 @@
   <script src="../blocks_vertical/motion.js"></script>
   <script src="../blocks_vertical/looks.js"></script>
   <script src="../blocks_vertical/procedures.js"></script>
-  <script src="../blocks_vertical/operators.js"></script>
-  <script src="../blocks_vertical/sound.js"></script>
   <script src="../blocks_vertical/pen.js"></script>
+  <script src="../blocks_vertical/sound.js"></script>
+  <script src="../blocks_vertical/sensing.js"></script>
   <script src="../blocks_vertical/data.js"></script>
   <script src="../blocks_vertical/extensions.js"></script>
